Enhancement: support for versioning.
There might be common modules to multiple versions, but we can take care of that in later refactors. For now I am going to assume versions are separate packages.

In terms of publishing, npm packages will only bundle a single version at a time.
So when imported, users won't see the version numbers.
They will only know the version based on the package's version.
This is why we must support packaging of separate versions.