Enhancement: execution to create relationship with parent.
Execution is opposite of all models.
Whereas other models have methods to add a child, execution creates relationship with parents. In the end, it is one and the same,
but writing logic this way prevents a confusion, since execution needs to be parented by both deployment and environment. So the code would have look weird or was error prone if either would have failed to claim execution. Hence execution does it in one place.