Enhancement: actions now support inputs.
An input to an action is synonymous to a configuration value.
Inputs can be used to change the behavior of the action slightly,
such as passing an encryption flag, or applying different tags.
An action needs to define a set of input keys,
and users must provide a value for those keys.