Enhancement: OctoAws deprecated use of replication in ECR.
ECR replication can take up to 30 minutes, and in edge cases more than that.
This is a big delay that users cannot afford, and can bring several issues of its own,
mainly around pipelines stuck in retry loop trying to fetch ECR image that is waiting for replication.

Instead, we have enabled the images to be uploaded in every region.
The image is still a shared resource, meaning ECR in one region would
automatically be counted in all regions.
As soon as other region starts to use the image, it will come in model diff as an update.
In turn, it would trigger a resource diff, where it would most likely trigger an update diff as well.
The shared-resource diff relies on replication region field to check if the region already
has an image or not. Depending, ADD or DELETE diff is produced.
Furthermore, add-image resource specifically checks for presence of image in that region before
putting a new one.