Enhancement: Octo removed support for hooks.
Hooks are not a well fleshed thought.
In future, we might give the task to state provider to record all events from the system.
Or create a new notification system just to notify of all the events,
but having to react on changing something has not worked out.

E.g. adding an nginx execution on adding environment has too many assumptions,
like name of image, that we are not comfortable taking.