patch(feat): octo | model now defines context via setContext.
getContext() was always probing parents to get the context.
It was slow, and required access to parents all the time.
But during unSynth() access to parents is not there,
and so it got difficult to unSynth() using anchors.

For this reason, we now save `context`, and getContext()
is just a getter to this variable.
We can set this value during unSynth() and anchors can
use the value directly.