patch(fix): octo | loading of resources from previous state.
Note 1: Unlike overlays, resources needs to be deserialized twice, once for new resources
and once for old resources. Resources are generated as a product of model and overlay diff.
 The user repository code, when runs with a previous state, model and overlay diff won't generate new resources.
That equates to ResourceDataRepository being empty.
On resource diff with the previous resource, DELETE diffs will be produced.
Therefore, it's necessary to load the ResourceDataRepository with new before deserializing the resources.
Note 2: Post commit you can still reference the running overlays, but any resource manipulation should be done
on after fetching the resource from the ResourceDataRepository. Never use the running resources because
post commit they are replaced by a new copy.