patch(fix): octo-aws-cdk | update based on value not working.
In certain cases, we use a manifest object to hold update values.
We pass these objects directly to the resource actions,
but also delete/clear those objects in the next line.
So instead, we need to pass a copy to the diff
so deletion in diff() does not affect them.