patch(fix): octo | new factories must be registered as default.
If a bunch of Container.get() requests are made for a factory,
and that factory is not registered yet, then the first get() method
must register that factory as default.
It also makes sense because there are no other factory registered,
and the get() is called with/without any metadata,
then there is no reason for that factory to not be a default.

If the default is not set to true,
then every get() method will generate a new promise for the same factory,
and there will be duplicate factories registered.