patch(feat): octo | initialize() now supports initialization of decorated entities.
In TypeScript, if a class or method is decorated, the decorator won't instantiate
unless the file was imported. In a repository, this might not be totally obvious
to import. Even if it is known, and users have to manually do something like,
`import './my-module.js'` then it looks ugly.

The initialize method now provides an option to pass Constructable definitions.
This not only enforces you to import the file, but also use that value declaratively.

You can pass things like Modules, EventListeners, and more.