patch(feat): octo | event listeners now receive past events replayed to them.
The original method is wrapped in another promise,
where first the instance of listener is get.
Without fetching the instance of listener, the `this` context is lost in the listener method.
Then, the events are replayed.

Note: Octo does not stop for events to be processed.
Its a fire and forget call. But it does attach a catch() to avoid unhandled promise rejections.
Ideally, the listener authors will handle their own methods.