patch(feat): octo | Module loading is now declarative.
Previously, modules were auto loaded just because a @Module() decorator was found.
But in reality, its difficult to visualize what is loaded.
Additionally, testing was difficult when we want to selectively load modules.

A better approach is to declaratively load only the modules we want.
So Octo now exposes `loadModules()` where you express the desire to load a module.
By default, all modules are registered, but they are hidden.
Only on load, they get un-hidden.
Same in test, you must declare the modules you want to load before test runs.