chore(feat): octo | Transaction Scenarios UT.
These tests contain different scenarios of a transaction.
It takes into account that previous transaction might have resulted in an error,
and that might have put resources in a dirty state.

Throughout the tests, some scenarios demonstrate recovery of state,
others genuinely fail when there are no ways to safely move to the desired state.