patch(fix): octo | duplicate resource processing.
- Also removes yield on dirty diffs from transaction.

We realized that we were processing the same diffs on both new-old and new-actual.
We instead need to create two separate diffs - new diffs vs dirty diffs,
where new diffs are changes that should have been applied if all was good,
and dirty diffs are changes that is on top of new to sync resources to desired state.

Here is how it works,
- New diffs are generated.
- Actual diffs are generated.
- Actual can be more, less, or equal to the new diffs, depending on revert, fix-forward, or some other scenarios.
- So we first process new. It should be a subset of actual. We should remove diffs that actual doesn't think we need to do.
- Then the remainder is processed as dirty diffs.