patch(feat): octo | OverlayDataRepository factory with initializing new overlays.
We had previously removed the ability to initialize OverlayDataRepository with a predefined overlays.
This was because users no longer need to "restore" previous state.
However, Octo's future UI might still want the ability to restore the previous state to read data.

Now we have the ability to initialize, but we don't expect users to ever use it.
Moreover, when re-initialized, the "newOverlays" will just be loaded with deserialized overlays.
Previously the same data was kept as "oldOverlays" and "newOverlays" was reserved for new changes.
But since we don't expect to make any new changes upon re-initialization, we did not consider keeping old and new variables both.