patch(fix): octo | Overlay unSynth() to add anchors without dependencies.
When overlays are deserialized, anchors are added, and if the same `addAnchor()` method
is used, then it would require all dependencies to be present. However, the dependencies
are deserialized later, and so the code would fail.

Instead, we are now unSynth the overlay by directly pushing anchors to it.
Later, when dependencies are deserialized, the whole dependency connection would take place,
and everything will be set properly.

And as previous commit stated, deserialization of models are overlays is not something the user
will ever do, so this operation seems to be safe.