patch(feat): octo | TestContainer bootstrap.
The TestContainer now has the ability to bootstrap the container with all its internal factories.

We decided to expose Container to our users.
That means there is a chance they might want to create TestContainer themselves for their tests.
That means they will need to bootstrap the container, but we can't let them do that without exposing the whole internal structure.
That is why the TestContainer will bootstrap on every load.

And that has other benefit too,
- our own test cases are now simpler, since they don't have to mock a bunch of things.
- A pristine container is loaded before every test.