patch(fix): octo | keep the same container instance between tests.
The container instance, if created using "forceNew",
will be different from the container instances referenced by decorators.

This is because the container instances in decorator is doing
Container.getInstance(), but TestContainer then re-initializes the
container using Container.getInstance(true).
This generates a new container instance which is different from the
container instance held by the decorators.

As a solution, we are now keeping the same container instance alive
throughout the tests within the same thread.