patch(fix): octo | module-container reset must use a sync for-of loop.
Previously using map(), the modules were not being reset properly
since map() for some reason was behaving async and the reset()
method was returning before the map had a chance to work through all its callbacks.

Using a sync for-of loop we ensure that all modules are reset before the
reset() method returns.