patch(feat): octo-aws-cdk | resources are now independent of other resources.
A resource that is expected to work independently, i.e. the resource
can take any other resource as its parent, not just under the @octo package,
must not rely on @octo package resources.

For this reason, all interdependency is now defined as separate types.
The resource expects a resource, and using types assumes that the resource passed is of that type.

Clearly, this is not an enforced solution, but resource creators are free to put more checks in the constructor.