patch(feat): octo-aws-cdk | Resources with parent check.
In a module world, different resources needs to interact with other resources.
They are not sure what the parents can be, but they can define a schema that they expect.
But in schema there is no way to ensure that the parent is of certain type.

This is also because resource type is not standardized. One resource owner can name something vpc and other can name is aws-vpc.
For this reason, the resources needs to do a type check before they accept the parent.

Over time, I do expect people to standardize the names based off of our documentation, but nothing more we can enforce.

One final note, to reiterate,
- Resources have their own schema. Putting @Validate() in them will work.
- Resources accepts parents. But they cannot import the parent schema from anywhere. So they define their own. These schemas are only TS types. Putting @Validate() in them is not going to work because the values are never set as per this schema. The values are given to this resource, which will never trigger this schema's validation rules.
- The parents of resources are only relying on TS type to hint the module on what its supposed to pass.
- This is the reason why modules also have to define a schema for resources.