patch(feat): octo | module inputs now take true arguments.
Previously we were forcing users to pass only strings when providing
inputs for modules. This was done to ensure that models or resources
can be passed as references that were strings.

That was causing a lot of confusion with other non-string inputs,
like objects. So now the module inputs will be the true type,
and now instead we release stub() which is mimics the type when
provided a string.