patch(feat): octo-aws-cdk | module resource inputs deprecated in favor of resource derivation.
We no longer prefer to ask for resources as inputs, but rather want module owners to derive the resources by asking octo to match a resource using a schema they provide.

This kind of derivation allows users to be free of keeping track of resources and many other inputs, such as aws-region-id.

Of course the schema that module provides might not exactly match the schema of the resources in the system, but that is when octo will report the failure and users are expected to write a schema translation.

We are consciously putting this burden on users as they are the only one who can write this translation to make their infra work.
This translation comforts the modules to freely work with any schema as users will provide a translation.

Schema drifting is inevitable, and schema translation is going to play a larger role when using 3rd party modules.