patch(feat): octo | ability to search for resources using models.
The method getResourceMatchingSchema() returns the synth() of resource and the resource itself.
The synth(), if translated, will contain properties and responses that differ from the resource, because that's been translated.

Module owners must understand when to use which. To read data, use the synth(), vs to use the resource, like adding as parent, use the resource.