patch(feat): octo | support to search model with schema.
Octo now supports searching for models based on schema.
Previously, modules were relying on the model graph to look for parent models, but that is not true, as the model graph may be different than what the modules are expecting.
A better way is to even search models and overlays using schemas.