patch(fix): octo-aws-cdk | s3-storage resource does not update.
The S3 storage resource should not support uploading of files.
Its responsibility is only to create a S3 bucket for storage.
Supporting files in S3 is a product in itself, and that is not
what Octo is setup to do. Its an infrastructure, so it should only
create infrastructure.

S3 website is a different concept, as it manages the website.
S3 storage is not the same as that.
S3 storage is more similar to ECR in this regard.