patch(fix): octo | test resources must copy resources in actual.
Previously, the same resource as new was being used for actual as well,
which is wrong, since any change to new will automatically make that
change to actual, as they are shared by reference.
This caused the side effect of no diff on change, because on changing the
new resource, the actual already had the change.

As a solution, we are now always cloning the new resource before placing that resource in actual.

This created an additional bug was during commit,
where if multiple resources are being committed out of order,
then deReference wouldn't find them.
For that reason we created the utility that sorts the resources as per their dependency.

But that same sorting won't be applied on methods of test-resources,
since it is creating the resources on the spot. So we expect the
deReference method to always find the parents when searching in actual resource repository.