patch(fix): octo | resource unsynth to resolve MatchingResource parents.
When user sets matching parents on a resource, they expect that matching
resource to be received back. In delete diff, after unsynth, the users
are receiving the old resource, and so unsynth needs to put the exact same
matching resource that the user originally gave.

This is not a problem for anchors because users never get to see the old
anchors, just in resource they can get the old resource back in delete diff.
That is why the unsynth needs to take care of it.