patch(fix): octo | cloning of parent resource was duplicating.
The new parent was being pushed to self parents array 2 times,
one from addChild() and one from push.

When we fixed that we realized that the parents needs to have the same
structure - i.e. matching vs non-matching because when resources are
unSynthed, then users get to access the old resource parents, and if
they passed a parent as matching, they still expect them to be matching.