patch(fix): octo | calculate schema on-demand.
Sometimes the node schema can be stale, because the source of the schema
comes from synth(), esp in case of resources, where response might not have
been present when synth() was called the first time and recorded.

If not fetched on-demand, there is a risk of not providing the most current schema.
Thus, we should always calculate schema on-demand.