patch(fix): octo-aws-cdk | s3-storage behavior not copied.
S3-Storage parent is a matching resource which was only added using addPermission() method.
However, when the actual resource was being created in transaction-service, unSynth() was being called.
And unSynth() only adds the parent using constructor.

This ended up only adding parent relationship, but was not copying other
behaviors in their dependency as being done in addPermission().
And since addPermission() is never called in the whole unSynth() chain, these dependency behaviors never got copied in actual resources, but were present in the old resources.

The solution was to follow the pattern, like in ecs-task-definition, or ecs-service where a separate method is dedicated to add the behaviors.
Then this method is called in the constructor, which ensures that during unSynth() all behaviors are also copied.

Cons,
- The way S3-Storage constructor is defined, users need to know to never
add the parent using constructor but to always use the addPermission() method. The call from constructor is only used by unSynth().
- This is a bit of discrepancy from other patterns, but we can work to consolidate the logic later.