patch(fix): octo-aws-cdk | iam-role to delete all policies upon deletion.
Previously, the delete diff was divided into 2 parts - update to iam role
first to delete all policies, and then the delete iam role diff was processed.
This led to situations where iam-role policies were deleted first, but before the iam-role could be deleted, the transaction was getting failed due to several reasons, such as on internet gateway.
So, we moved all deletion of policies in the delete role itself, and let go of all the update policy, i.e. if an iam-role is being deleted, it incorporates all deletion logic within itself.
The update policies are only effective when a single policy is being updated.