patch(fix): octo | resources can be force deleted.
In resources, there are situations when a node needs to be deleted even if there are children nodes present. E.g. iam-role needs to be deleted when s3-storage exists.

Without force, the iam-role cannot be deleted.
New resources won't have iam-role, and actual resources will have.
So a delete diff will be generated, and actual resource will try to be deleted, which is when error was happening.

For resources, there is nothing that can delete the child relationship.
So a force option is required here.