patch(fix): octo-aws-cdk | incremental diff inverse.
If there are multiple diff updates, each update focusing on a specific part of property, then diffInverse() should not clone all properties.
E.g. iam-role is being deleted, and it generates 3 diff updates to first delete the 3 policies it has. The first delete when processed, will generate a diffInverse() which would have deleted all properties, which is wrong.
Instead, now they only delete the one property for which the diff is about.