patch(feat): octo | deprecate shared-resource.
Concept of shared-resource is redundant.
Users can simply define a resource with the same resourceId,
and that would be equivalent to a shared-resource.

To create a shared-resource, create a resource with the same resourceId
as before. When the same resource is encountered by transaction-service
applyModels(), it would then invoke the merge() function to copy all properties and response from current to previous.
It would also ensure that the reference to current also points to previous, so that any usage of current is same as using previous.