patch(feat): octo | no duplicate nodes.
In octo we shouldn't allow duplicate nodes to exist.
Duplicate nodes can be created in modules and its handles.
These nodes when start to create relationships using addChild(),
are handled by skipping the creation of dependencies.

Furthermore, data repositories now also throw error when they see duplicate nodes.

We had to modify setContext() so that we can deprecate the use of Object.is() in addChild() and addFieldDependency() methods. Use of is() method was allowing duplicate nodes to exist because duplicate nodes are technically different objects, so they were not caught using is() method.

Resource's parent list are now also checking before adding the new parent. Changes to MatchingResource are also for the same reason.