patch(feat): octo-aws-cdk | response schema should be optional.
The response schema, by its nature, must all be optional.
They are only available when used in resource actions.
So we introduced a new method getSchemaInstanceInResourceAction() whose type is set to give required response back.
In other parts of the code, we must use getSchemaInstance(), but when in resource actions accessing response, we should use the new method.

Its just to makes the types better, and does not affect functionality.