patch(feat): octo-aws-cdk | restrict overlay schema export.
Model and Overlays are internals of modules, and they should not be exported. That includes not exporting their schemas as well.
Previously, we used to create dependencies between overlays of different modules, but that is not required. Instead, create an anchor whose parent is overlay, and the anchor is present in the model.

That allows the anchor to still be found using model `model.getMatchingAnchor()`, and the parent of the anchor is an overlay,
so any other overlay depending on this anchor will automatically create the other anchor as its parent, because that's what overlays do with anchors - they create dependencies with the parent of the anchor.

This is a pretty twisted, but great use of anchors - to have parent different than where the anchor lives. This is exactly how octo envisioned it. An anchor can be placed anywhere, but its parent can be it itself, or its overlay.