patch(fix): octo-build | fix node_modules eslint error.
Since there are multiple package.json files, the npm package manager
decides which packages go where. It will place the dependency in root
node_modules if it finds a package already there with same version,
otherwise it will place it in the package's node_modules.

Due to dependency being in multiple places, eslint gets confused
and throws error on not able to find yeoman or yargs packages.

In eslint, we are telling it to look into both node_modules.
And we are also telling eslint to ignore any node_modules files from
being linted.