patch(feat): octo | "log" method in abstract action is synchronous.
The log events will be emitted, and its a fire-and-forget call.
Even with async it was like this, because logs emit events,
and events are handled in the next callback execution cycle.

With this method being sync, there will be 2 callback execution cycle when the log will actually show up.

Most of the time, it should be fine since eventService will have been resolved. But if we observe too much lag on logs, we will make this async again.