patch(fix): octo | Missing events!
Many events, specifically registration events, were missing!
This was because during event-source decorator,
we had a promise, which was causing a race condition between
transaction service being created and decorator being applied.

As a result, all registration events were running without the decorator
being applied, since the decorator was being applied on another instance
which was not under Container.