patch(feat): octo | resource action mock enhancements.
There are several changes in this PR.
- Resource action must return a response, rather than setting it inline.
This allows a better interface where a set type is returned.
On the Octo library, we set this response.
- mock() method should not really mock the aws clients.
Instead, they should just set the response.
The idea of the mock() is to do the same operation as handle(),
but without any actual aws calls.
- Generic resource tagging doesn't implement a mock method
since tag mocks don't change response.
Instead, every resource mock for tags just return the response as is.
The handle() method also just returns any to ensure resource action handle() methods can return the right type. This can be improved, but its fine for now.

Ultimately, the idea with mock is to allow capturing of a resource state.
On a resource action level, that simply means ignoring AWS calls,
since users are expected to have done the resource change themselves.
We are just updating the state of Octo.

For add, users have created the resource, and updated the code.
Response is set as per capture.
For update, users have already updated the resource, and updated the code.
Response is set as per capture.
For delete, users have already deleted the resource, and updated the code.
Response is set as per capture. But since the resource is being deleted,
this has no meaning as the resource state will be deleted.
But its good for consistency to return response.

Caution: its really a caution for users using capture.
Its easy to capture something wrong, since response setting is not checked.
We will just set whatever the user tells us.
The code might be pointing to a certain state, but has capture pointing to
something entirely different.
With time we might be able to apply constraints, but for now this is ok.