patch(fix): octo-aws-cdk | account modules must register aws factories.
If only app and account module is to be loaded,
then the application would crash because the module tries to get
STSClient from container, but there are no imports that registers
the factory.

That is because the only file for account modules that imports the
aws factories are the module itself, and it was importing only the type.

Thus, we need to ensure that the actual file is imported in order
for the factory to be registered.