patch(fix): octo-build | octo yaml config to separate imports from modules.
When module contains import as well, then when the module is deleted
to signify a delete operation, the imports are lost as well.
This effectively breaks the program because the import is needed to
register the necessary classes to unSynth().

Thus, the imports are separated out, so that even when module definitions are removed, the imports allows processing of those modules.