patch(feat): octo | register module hooks even if module not loaded.
Currently, if a module is not loaded, i.e. is imported,
but no module definition provided, the module is completely skipped.

However, the module might still want to run its hook for cleanup.
To support this, we have decided to load the module hooks as long as
it is imported. It does not matter if the module is loaded or not.

This allows for the hooks to always be present for a module.
Depending on the hook itself, it will run regardless of module's definition.
So it is important for hooks to be written in a way that it does not fail.