patch(feat): octo-aws-cdk | s3-website module post commit hook.
Just like the hook responsible for saving manifest state,
a post hook is created to remove the manifest state when the bucket
is being deleted.

The post and pre hooks, both, are written in a way to only apply
on certain conditions, and should not fail if those conditions
don't apply.