patch(feat): octo | support for app lock.
App lock is a feature in Octo that ensures that only one state update
happens at a time. In scenarios where there are multiple users
attempting to update the state, atomicity becomes an absolute requirement.

Imagine a case for pipeline where multiple PRs are attempting to
be merged. Only 1 PR needs to merge at a time, and each PR must have
the latest state.

More documentation will be noted down, but the overview is,
* Lock state separately, outside of Octo transaction.
* Call Octo with or without lock ID.
* If without, transaction won't initiate.
* If with, transaction will check if app is locked before executing transaction.
* Unlock the state separately, outside of Octo transaction.